app/main.py
from fastapi import FastAPI
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

app = FastAPI()

@app.get("/health")
async def health_check():
    return {"status": "ok"}