requirements.txt
fastapi
uvicorn
requests
schedule
firebase-admin
sqlalchemy
feedparser
pytest
python-dotenv